Split `definition-context.ts`
