+++ src/component/definition/definition-context.key.ts
@@ -1,0 +1,7 @@
+import { SingleContextKey } from 'context-values';
+import { DefinitionContext } from './definition-context';
+
+/**
+ * @internal
+ */
+export const definitionContextKey = new SingleContextKey<DefinitionContext<any>>('definition-context');
