+++ src/component/definition/index.ts
@@ -1,3 +1,4 @@
 export * from './component-factory';
 export * from './custom-elements';
 export * from './definition-context';
+export * from './element-base-class';
